refactor: 카드 번호 앱에서 분리
